+++ css/custom.css
@@ -557,18 +557,29 @@ canvas {
 
 
 /* services */
+
+.bg-services {
+    background-color: #F8F9FA;
+}
 .servicesjs {
+    
     display: inline-block;
-    width: 28%;
-    float: left;
     margin-top: 10px;
-    padding: 20px;
+    padding: 24px;
     text-align: center;
     border: 1px dotted #e7e7e7;
-    margin-right: 30px;
     }
 .servicesjs > p {
-    margin-top: 10px;
+    margin-top: 24px;
+    color: #6c757d;
+}
+.servicesjs > h5 {
+    font-size: 1.25rem;
+    margin-top: 40px;
+}
+.servicesjs > span {
+    margin-bottom: 1rem;
+    font-size: 40px;   
 }
 /* statistics */
     .statistics{
