+++ index.html
@@ -135,11 +135,9 @@
         </div>
     </div>
     <!-- services -->
-    <div class="container">
-        <div class="services row">
-            <div class="services-topas"></div>
-            <div class="services-bot"></div>
-            SERVICES
+    <div class="container bg-services">
+        <div class="row">
+            <div class="services"></div>
         </div>
     </div>
     <!-- statistics -->
