+++ js/components/services/renderSingleService.js
@@ -3,8 +3,8 @@
 function renderSingleService(services, i) {
 
     return `<div class ="servicesjs">
-                <img src="#" alt="#">
-                <h3>${services.header}</h3>
+                <span class=${services.icon}></span>
+                <h5 class="font-weight-bold">${services.header}</h5>
                 <p>${services.paragraph}</p></div>
 </div>`
 }
